chore(store.ts): use configureStore from @reduxjs/toolkit
* added missing types, schemas separated, array.includes, async/await
