+++ package.json
@@ -9,9 +9,12 @@
     "@redux-devtools/dock-monitor": "^4.0.0",
     "@redux-devtools/log-monitor": "^5.0.0",
     "@redux-devtools/slider-monitor": "^5.0.0",
+    "@types/humps": "^2.0.6",
+    "@types/lodash": "^4.14.202",
     "@types/node": "^20.11.20",
     "@types/react": "^18.2.58",
     "@types/react-dom": "^18.2.19",
+    "@types/redux-logger": "^3.0.13",
     "@typescript-eslint/eslint-plugin": "^7.0.2",
     "@typescript-eslint/parser": "^7.0.2",
     "@vitejs/plugin-react-swc": "^3.6.0",
