--- src/store/configureStore.ts
@@ -1,20 +1,0 @@
-// eslint-disable-next-line @typescript-eslint/ban-ts-comment
-// @ts-nocheck
-import { createStore, applyMiddleware, compose } from "redux";
-import { thunk } from "redux-thunk";
-import { createLogger } from "redux-logger";
-import api from "../middleware/api";
-import rootReducer from "../reducers";
-import DevTools from "../containers/DevTools";
-
-const configureStore = (preloadedState) => {
-  const store = createStore(
-    rootReducer,
-    preloadedState,
-    compose(applyMiddleware(thunk, api, createLogger()), DevTools.instrument()),
-  );
-
-  return store;
-};
-
-export default configureStore;
