--- src/types/index.ts
@@ -1,10 +1,0 @@
-export type CamelCase<S extends string> =
-  S extends `${infer P1}_${infer P2}${infer P3}`
-    ? `${Lowercase<P1>}${Uppercase<P2>}${CamelCase<P3>}`
-    : Lowercase<S>;
-
-export type KeysToCamelCase<T> = {
-  [K in keyof T as CamelCase<string & K>]: T[K] extends Record<string, never>
-    ? KeysToCamelCase<T[K]>
-    : T[K];
-};
